requirements-dev.txt
pytest==5.3.5
black==19.10b0
isort==4.3.21-2
grpcio==1.26.0
grpcio-tools==1.26.0
